PFDA 3786 jobs syncing (#2394)
* PFDA-3786 first routes

* PFDA-3786 switching context

* PFDA-3786 WIP

* PFDA-3786 added preparation for output files syncing

* PFDA-3786 getting list of output files

* PFDA-3786 added synchronization with frontend

* PFDA-3786 added file event

* PFDA-3786 little refactor and one missing feature

* PFDA-3786 added more tests

* PFDA-3786 added more tests

* PFDA-3786 small fixes for data generation

* PFDA-3786 fix describe handling and error for items key

* PFDA-3786 added missing functionality

* PFDA-3786 added challenge check job

* PFDA-3786 getting rid of old job_sync code

* PFDA-3786 readability refactor

* PFDA-3786 forgotten stuff

* PFDA-3786 fix tests

* PFDA-3786 fix test

* PFDA-3786 removed unused test

* PFDA-3786 changed import

* PFDA-3786 added config

* PFDA-3786 fix compilation issues

* PFDA-3786 removed forgotten things

* PFDA-3786 checkChallenge enabled by default

* PFDA-3786 optimalization

* PFDA-3786 disabled challenge job syncing for tests

* PFDA-3786 removed checkChallengeJobs

* PFDA-3786 if challenge user is not found the synchronization is not started

* PFDA-3786 switched to bulk find method

* PFDA-3786 addressed Ludvik's comments

* PFDA-3786 changed status text

* PFDA-3786 added comment

* PFDA-3786 refactor

* PFDA-3786 fixed config

* PFDA-3786 switched back to main queue

* PFDA-3786 fixed removing of finished task

* PFDA-3786 removed unused import

* PFDA-3786 fix

* PFDA-3786 refactor

* PFDA-3786 fix tests

* PFDA-3786 add removed duplicit

* PFDA-3786 reformat

* PFDA-3786 fixed typing in Notification

* PFDA-3786 fixed typing

* PFDA-3786 added generics to BullJob

* PFDA-3786 typing refactor

* PFDA-3786 change in usage of EntityManager

* PFDA-3786 switched handler functions into handler classes

* PFDA-3786 fix bad import

* PFDA-3786 refactor towards greater generality

* PFDA-3786 fix running in a space

* Fix import

* PFDA-3786 addressing comments

* PFDA-3786 refactored

* PFDA-3786 refactor

* PFDA-3786 last set of tests and linting

* PFDA-3786 removed usage of a type that will come in a different PR :-D

* PFDA-3786 fix import

* PFDA-3786 import fix

* PFDA-3786 import change

* PFDA-3786 temporarily add debug logging

* PFDA-3786 change in logger

* PFDA-3786 just little updates

* PFDA-3786 added job sync to Workflows

* PFDA-3786 added page refresh on AppExecutions

* PFDA-3786 fix test

* PFDA-3786 added support for terminating state as well

* PFDA-3786 removed logging used for debugging

---------

+++ https-apps-api/packages/shared/src/domain/user-file/asset.entity.ts
@@ -45,7 +45,7 @@ class Asset extends Node implements IFileOrAsset, ITrackable {
   @Property()
   scope: string
 
-  @Property({ type: 'bigint' })
+  @Property({ type: 'numeric' })
   fileSize?: number
 
   // unused FK references
